renamed login package to authenticate
